Fixes serialization
